fast_api/pyproject.toml
[tool.poetry]
name = "fast-api"
version = "0.1.0"
description = ""
authors = ["Moivvas <moivvas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
psycopg2 = "^2.9.9"
jinja2 = "^3.1.2"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
libgravatar = "^1.0.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
fastapi = "^0.104.1"
pydantic = {extras = ["email"], version = "^2.5.1"}
pydantic-settings = "^2.1.0"
slowapi = "^0.1.8"
cloudinary = "^1.36.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
